Split chat reducer slices into files
