+++ src/components/chat/chat-reducer.ts
@@ -1,1670 +1,1 @@
-/**
- * Chat state reducer for managing chat UI state.
- *
- * This reducer handles all state transitions for the chat interface:
- * - WebSocket message handling
- * - Session management (loading, switching)
- * - Permission and question requests
- * - Tool input streaming updates
- *
- * The state is designed to be used with useReducer for predictable updates.
- */
-
-import type {
-  ActiveHookInfo,
-  ChatMessage,
-  ChatSettings,
-  ClaudeMessage,
-  CommandInfo,
-  MessageAttachment,
-  PendingInteractiveRequest,
-  PermissionRequest,
-  QueuedMessage,
-  SessionInfo,
-  SessionInitData,
-  SessionStatus as SharedSessionStatus,
-  TokenStats,
-  ToolUseContent,
-  UserQuestionRequest,
-  WebSocketMessage,
-} from '@/lib/claude-types';
-import {
-  createEmptyTokenStats,
-  DEFAULT_CHAT_SETTINGS,
-  getToolUseIdFromEvent,
-  isStreamEventMessage,
-  isWsClaudeMessage,
-  MessageState,
-  updateTokenStatsFromResult,
-} from '@/lib/claude-types';
-import { createDebugLogger } from '@/lib/debug';
-
-// Debug logger for chat reducer - set to true during development to see ignored state transitions
-const DEBUG_CHAT_REDUCER = false;
-const debug = createDebugLogger(DEBUG_CHAT_REDUCER);
-
-// =============================================================================
-// State Types
-// =============================================================================
-
-/** Information about a rejected message for recovery */
-export interface RejectedMessageInfo {
-  text: string;
-  attachments?: MessageAttachment[];
-  error: string;
-}
-
-/** Content stored for a pending message (for recovery on rejection) */
-export interface PendingMessageContent {
-  text: string;
-  attachments?: MessageAttachment[];
-}
-
-/**
- * Pending interactive request - a discriminated union that makes it impossible
- * to have both a permission request and question request simultaneously.
- * Replaces separate pendingPermission and pendingQuestion nullable fields.
- */
-export type PendingRequest =
-  | { type: 'none' }
-  | { type: 'permission'; request: PermissionRequest }
-  | { type: 'question'; request: UserQuestionRequest };
-
-/**
- * Session status - a discriminated union that makes invalid states unrepresentable.
- * Replaces separate running, stopping, loadingSession, and startingSession booleans.
- *
- * State transitions:
- *   idle → loading → starting → ready ↔ running → stopping → ready
- */
-export type SessionStatus =
-  | { phase: 'idle' }
-  | { phase: 'loading' }
-  | { phase: 'starting' }
-  | { phase: 'ready' }
-  | { phase: 'running' }
-  | { phase: 'stopping' };
-
-/** Tool progress tracking for long-running tool executions */
-export interface ToolProgressInfo {
-  toolName: string;
-  elapsedSeconds: number;
-}
-
-/** Task notification from SDK (e.g., Task tool subagent updates) */
-export interface TaskNotification {
-  id: string;
-  message: string;
-  timestamp: string;
-}
-
-/** Rewind preview state for displaying confirmation dialog */
-export interface RewindPreviewState {
-  /** User message UUID we're rewinding to */
-  userMessageId: string;
-  /** Unique identifier for this specific rewind request (for race condition handling) */
-  requestNonce: string;
-  /** Whether we're currently loading the preview or executing the rewind */
-  isLoading: boolean;
-  /** Whether the actual rewind is in progress (vs just previewing) */
-  isExecuting?: boolean;
-  /** Files that would be affected (populated after dry run completes) */
-  affectedFiles?: string[];
-  /** Error message if preview failed */
-  error?: string;
-}
-
-export interface ChatState {
-  /** Chat messages in the conversation */
-  messages: ChatMessage[];
-  /**
-   * Session lifecycle status - a discriminated union that makes invalid states unrepresentable.
-   * Replaces separate running, stopping, loadingSession, and startingSession booleans.
-   */
-  sessionStatus: SessionStatus;
-  /** Current git branch for the session */
-  gitBranch: string | null;
-  /** Available Claude CLI sessions */
-  availableSessions: SessionInfo[];
-  /**
-   * Pending interactive request awaiting user response.
-   * Discriminated union ensures only one request type can be active at a time.
-   */
-  pendingRequest: PendingRequest;
-  /** Chat settings (model, thinking, plan mode) */
-  chatSettings: ChatSettings;
-  /**
-   * Queued messages waiting to be sent.
-   * Map from message ID to QueuedMessage - enforces uniqueness by design.
-   * Maps automatically de-dupe: adding the same ID twice simply overwrites.
-   */
-  queuedMessages: Map<string, QueuedMessage>;
-  /** Tool use ID to message index map for O(1) updates */
-  toolUseIdToIndex: Map<string, number>;
-  /** Latest accumulated thinking content from extended thinking mode */
-  latestThinking: string | null;
-  /**
-   * Pending messages awaiting backend confirmation (shown with "sending..." indicator).
-   * Map from message ID to content. Presence in map = pending, content = for recovery on rejection.
-   * Replaces separate pendingMessageIds Set and pendingMessageContent Map.
-   */
-  pendingMessages: Map<string, PendingMessageContent>;
-  /** Last rejected message for recovery (allows restoring to input) */
-  lastRejectedMessage: RejectedMessageInfo | null;
-  /** Whether context compaction is in progress (placeholder for future SDK support) */
-  isCompacting: boolean;
-  /** Tool progress tracking for long-running tools - Map from tool_use_id to progress info */
-  toolProgress: Map<string, ToolProgressInfo>;
-  /** Session initialization data (tools, model, etc.) from system init message */
-  sessionInitData: SessionInitData | null;
-  /** Current permission mode from status updates */
-  permissionMode: string | null;
-  /** Whether a compact boundary has been encountered in this session */
-  hasCompactBoundary: boolean;
-  /** Active hooks currently executing - Map from hook_id to hook info */
-  activeHooks: Map<string, ActiveHookInfo>;
-  /** Task notifications from SDK (e.g., subagent updates) */
-  taskNotifications: TaskNotification[];
-  /** Slash commands from CLI initialize response */
-  slashCommands: CommandInfo[];
-  /** Whether slash commands have finished loading for this session */
-  slashCommandsLoaded: boolean;
-  /** Accumulated token usage stats for the session */
-  tokenStats: TokenStats;
-  /**
-   * Queue of SDK-assigned UUIDs waiting to be mapped to user messages.
-   * Used when UUIDs arrive before their corresponding messages.
-   */
-  pendingUserMessageUuids: string[];
-  /**
-   * Map from user message ID (stable identifier) to SDK-assigned UUID.
-   * Populated as user_message_uuid events are received.
-   * Using message IDs instead of indices ensures correct mapping even when
-   * messages are inserted or removed from the array.
-   */
-  messageIdToUuid: Map<string, string>;
-  /**
-   * Set of user message IDs that were sent in the current session (not loaded from snapshot).
-   * Used to ensure UUIDs are only mapped to messages sent in this session,
-   * not historical messages from loaded sessions.
-   */
-  localUserMessageIds: Set<string>;
-  /** Current rewind preview state (null when not showing rewind dialog) */
-  rewindPreview: RewindPreviewState | null;
-}
-
-// =============================================================================
-// Action Types
-// =============================================================================
-
-export type ChatAction =
-  // WebSocket message actions
-  | { type: 'WS_STATUS'; payload: { running: boolean } }
-  | { type: 'WS_STARTING' }
-  | { type: 'WS_STARTED' }
-  | { type: 'WS_STOPPED' }
-  | { type: 'WS_CLAUDE_MESSAGE'; payload: { message: ClaudeMessage; order: number } }
-  | { type: 'WS_ERROR'; payload: { message: string } }
-  | { type: 'WS_SESSIONS'; payload: { sessions: SessionInfo[] } }
-  | { type: 'WS_PERMISSION_REQUEST'; payload: PermissionRequest }
-  | { type: 'WS_USER_QUESTION'; payload: UserQuestionRequest }
-  | { type: 'WS_PERMISSION_CANCELLED'; payload: { requestId: string } }
-  // Session actions
-  | { type: 'SESSION_SWITCH_START' }
-  | { type: 'SESSION_LOADING_START' }
-  // Tool input streaming action
-  | { type: 'TOOL_INPUT_UPDATE'; payload: { toolUseId: string; input: Record<string, unknown> } }
-  // Track tool use message index
-  | { type: 'TOOL_USE_INDEXED'; payload: { toolUseId: string; index: number } }
-  // Permission/question response actions
-  | { type: 'PERMISSION_RESPONSE'; payload: { allow: boolean } }
-  | { type: 'QUESTION_RESPONSE' }
-  // Stop action
-  | { type: 'STOP_REQUESTED' }
-  // User message action
-  | { type: 'USER_MESSAGE_SENT'; payload: ChatMessage }
-  // Queue actions (optimistic local state)
-  | { type: 'ADD_TO_QUEUE'; payload: QueuedMessage }
-  | {
-      type: 'MESSAGE_SENDING';
-      payload: { id: string; text: string; attachments?: MessageAttachment[] };
-    }
-  | { type: 'CLEAR_REJECTED_MESSAGE' }
-  // Message used as interactive response (clears pending request and adds message)
-  | { type: 'MESSAGE_USED_AS_RESPONSE'; payload: { id: string; text: string; order: number } }
-  // Settings action
-  | { type: 'UPDATE_SETTINGS'; payload: Partial<ChatSettings> }
-  | { type: 'SET_SETTINGS'; payload: ChatSettings }
-  // Thinking actions (extended thinking mode)
-  | { type: 'THINKING_DELTA'; payload: { thinking: string } }
-  | { type: 'THINKING_CLEAR' }
-  // Clear/reset actions
-  | { type: 'CLEAR_CHAT' }
-  | { type: 'RESET_FOR_SESSION_SWITCH' }
-  // Message state machine actions (primary protocol)
-  | {
-      type: 'MESSAGES_SNAPSHOT';
-      payload: {
-        /** Pre-built ChatMessages from backend - ready to use directly */
-        messages: ChatMessage[];
-        sessionStatus: SharedSessionStatus;
-        pendingInteractiveRequest?: PendingInteractiveRequest | null;
-      };
-    }
-  | {
-      type: 'MESSAGE_STATE_CHANGED';
-      payload: {
-        id: string;
-        newState: MessageState;
-        queuePosition?: number;
-        errorMessage?: string;
-        // For ACCEPTED state, includes full message content so we can add it to the list
-        userMessage?: {
-          text: string;
-          timestamp: string;
-          attachments?: MessageAttachment[];
-          settings?: ChatSettings;
-          /** Backend-assigned order for reliable sorting */
-          order: number;
-        };
-      };
-    }
-  // SDK message type actions
-  | { type: 'SDK_STATUS_UPDATE'; payload: { permissionMode?: string } }
-  | {
-      type: 'SDK_TOOL_PROGRESS';
-      payload: { toolUseId: string; toolName: string; elapsedSeconds: number };
-    }
-  | { type: 'SDK_TOOL_USE_SUMMARY'; payload: { summary?: string; precedingToolUseIds: string[] } }
-  | { type: 'SDK_TASK_NOTIFICATION'; payload: { message: string } }
-  | { type: 'SDK_COMPACTING_START' }
-  | { type: 'SDK_COMPACTING_END' }
-  // System subtype actions
-  | { type: 'SYSTEM_INIT'; payload: SessionInitData }
-  | { type: 'COMPACT_BOUNDARY' }
-  | {
-      type: 'HOOK_STARTED';
-      payload: { hookId: string; hookName: string; hookEvent: string };
-    }
-  | { type: 'HOOK_RESPONSE'; payload: { hookId: string } }
-  // Task notification management
-  | { type: 'DISMISS_TASK_NOTIFICATION'; payload: { id: string } }
-  | { type: 'CLEAR_TASK_NOTIFICATIONS' }
-  // Slash commands discovery
-  | { type: 'WS_SLASH_COMMANDS'; payload: { commands: CommandInfo[] } }
-  // User message UUID tracking (for rewind functionality)
-  | { type: 'USER_MESSAGE_UUID_RECEIVED'; payload: { uuid: string } }
-  // Rewind files actions
-  | { type: 'REWIND_PREVIEW_START'; payload: { userMessageId: string; requestNonce: string } }
-  | { type: 'REWIND_PREVIEW_SUCCESS'; payload: { affectedFiles: string[]; userMessageId?: string } }
-  | {
-      type: 'REWIND_PREVIEW_ERROR';
-      payload: { error: string; userMessageId?: string; requestNonce?: string };
-    }
-  | { type: 'REWIND_CANCEL' }
-  | { type: 'REWIND_EXECUTING' } // Actual rewind in progress
-  | { type: 'REWIND_SUCCESS'; payload: { userMessageId?: string } }; // Actual rewind completed
-
-// =============================================================================
-// Helper Functions
-// =============================================================================
-
-function generateMessageId(): string {
-  return `msg-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`;
-}
-
-function createClaudeMessage(message: ClaudeMessage, order: number): ChatMessage {
-  return {
-    id: generateMessageId(),
-    source: 'claude',
-    message,
-    timestamp: new Date().toISOString(),
-    order,
-  };
-}
-
-/**
- * Inserts a message into the messages array at the correct position based on order.
- * Uses binary search for O(log n) performance. Messages are sorted by their
- * backend-assigned order (lowest first).
- */
-function insertMessageByOrder(messages: ChatMessage[], newMessage: ChatMessage): ChatMessage[] {
-  const newOrder = newMessage.order;
-
-  // Binary search to find insertion point based on order
-  let low = 0;
-  let high = messages.length;
-
-  while (low < high) {
-    const mid = Math.floor((low + high) / 2);
-    if (messages[mid].order <= newOrder) {
-      low = mid + 1;
-    } else {
-      high = mid;
-    }
-  }
-
-  // Insert at the found position
-  const result = [...messages];
-  result.splice(low, 0, newMessage);
-  return result;
-}
-
-/**
- * Determines if a Claude message should be stored in state.
- * We filter out structural/delta events and only keep meaningful ones.
- */
-function shouldStoreMessage(claudeMsg: ClaudeMessage): boolean {
-  // User messages with tool_result content should be stored
-  if (claudeMsg.type === 'user') {
-    const content = claudeMsg.message?.content;
-    if (Array.isArray(content)) {
-      return content.some(
-        (item) =>
-          typeof item === 'object' && item !== null && 'type' in item && item.type === 'tool_result'
-      );
-    }
-    return false;
-  }
-
-  // Result messages are always stored
-  if (claudeMsg.type === 'result') {
-    return true;
-  }
-
-  // For stream events, only store meaningful ones
-  if (!isStreamEventMessage(claudeMsg)) {
-    return true;
-  }
-
-  const event = claudeMsg.event;
-
-  // Only store content_block_start for tool_use, tool_result, and thinking
-  if (event.type === 'content_block_start') {
-    const blockType = event.content_block.type;
-    return blockType === 'tool_use' || blockType === 'tool_result' || blockType === 'thinking';
-  }
-
-  // Skip all other stream events
-  return false;
-}
-
-/**
- * Checks if a message is a tool_use message with the given ID.
- */
-function isToolUseMessageWithId(msg: ChatMessage, toolUseId: string): boolean {
-  if (msg.source !== 'claude' || !msg.message) {
-    return false;
-  }
-  const claudeMsg = msg.message;
-  if (!isStreamEventMessage(claudeMsg)) {
-    return false;
-  }
-  const event = claudeMsg.event;
-  if (event.type !== 'content_block_start' || event.content_block.type !== 'tool_use') {
-    return false;
-  }
-  const block = event.content_block as ToolUseContent;
-  return block.id === toolUseId;
-}
-
-/**
- * Gets the tool use ID from a Claude message if it's a tool_use start event.
- */
-function getToolUseIdFromMessage(claudeMsg: ClaudeMessage): string | null {
-  if (!isStreamEventMessage(claudeMsg)) {
-    return null;
-  }
-  return getToolUseIdFromEvent(claudeMsg.event);
-}
-
-// =============================================================================
-// Initial State
-// =============================================================================
-
-/**
- * Creates the base reset state used by CLEAR_CHAT, RESET_FOR_SESSION_SWITCH,
- * and SESSION_SWITCH_START. This eliminates duplication and ensures all reset
- * actions clear the same fields consistently.
- */
-function createBaseResetState(): Pick<
-  ChatState,
-  | 'messages'
-  | 'gitBranch'
-  | 'pendingRequest'
-  | 'toolUseIdToIndex'
-  | 'latestThinking'
-  | 'pendingMessages'
-  | 'lastRejectedMessage'
-  | 'isCompacting'
-  | 'toolProgress'
-  | 'sessionInitData'
-  | 'permissionMode'
-  | 'hasCompactBoundary'
-  | 'activeHooks'
-  | 'taskNotifications'
-  | 'tokenStats'
-  | 'pendingUserMessageUuids'
-  | 'messageIdToUuid'
-  | 'localUserMessageIds'
-  | 'rewindPreview'
-> {
-  // Note: slashCommands is intentionally NOT reset here.
-  // - For CLEAR_CHAT: Commands persist because we're clearing messages in the same session.
-  // - For session switches: MESSAGES_SNAPSHOT handles clearing slashCommands separately,
-  //   and the backend will replay the stored slash_commands event for the new session.
-  return {
-    messages: [],
-    gitBranch: null,
-    pendingRequest: { type: 'none' },
-    toolUseIdToIndex: new Map(),
-    latestThinking: null,
-    pendingMessages: new Map(),
-    lastRejectedMessage: null,
-    isCompacting: false,
-    toolProgress: new Map(),
-    sessionInitData: null,
-    permissionMode: null,
-    hasCompactBoundary: false,
-    activeHooks: new Map(),
-    taskNotifications: [],
-    tokenStats: createEmptyTokenStats(),
-    pendingUserMessageUuids: [],
-    messageIdToUuid: new Map(),
-    localUserMessageIds: new Set(),
-    rewindPreview: null,
-  };
-}
-
-/**
- * Creates extended reset state for session switches, which also clears
- * queue, session status, and slashCommands (since different sessions may have different commands).
- */
-function createSessionSwitchResetState(): Pick<
-  ChatState,
-  | 'messages'
-  | 'gitBranch'
-  | 'pendingRequest'
-  | 'toolUseIdToIndex'
-  | 'latestThinking'
-  | 'pendingMessages'
-  | 'lastRejectedMessage'
-  | 'queuedMessages'
-  | 'sessionStatus'
-  | 'isCompacting'
-  | 'toolProgress'
-  | 'sessionInitData'
-  | 'permissionMode'
-  | 'hasCompactBoundary'
-  | 'activeHooks'
-  | 'taskNotifications'
-  | 'slashCommands'
-  | 'slashCommandsLoaded'
-  | 'tokenStats'
-  | 'pendingUserMessageUuids'
-  | 'messageIdToUuid'
-  | 'localUserMessageIds'
-  | 'rewindPreview'
-> {
-  return {
-    ...createBaseResetState(),
-    queuedMessages: new Map(),
-    sessionStatus: { phase: 'loading' },
-    slashCommands: [], // Clear for new session - will be sent when Claude starts
-    slashCommandsLoaded: false,
-  };
-}
-
-export function createInitialChatState(overrides?: Partial<ChatState>): ChatState {
-  return {
-    messages: [],
-    sessionStatus: { phase: 'loading' },
-    gitBranch: null,
-    availableSessions: [],
-    pendingRequest: { type: 'none' },
-    chatSettings: DEFAULT_CHAT_SETTINGS,
-    queuedMessages: new Map(),
-    toolUseIdToIndex: new Map(),
-    latestThinking: null,
-    pendingMessages: new Map(),
-    lastRejectedMessage: null,
-    isCompacting: false,
-    toolProgress: new Map(),
-    sessionInitData: null,
-    permissionMode: null,
-    hasCompactBoundary: false,
-    activeHooks: new Map(),
-    taskNotifications: [],
-    slashCommands: [],
-    slashCommandsLoaded: false,
-    tokenStats: createEmptyTokenStats(),
-    pendingUserMessageUuids: [],
-    messageIdToUuid: new Map(),
-    localUserMessageIds: new Set(),
-    rewindPreview: null,
-    ...overrides,
-  };
-}
-
-// =============================================================================
-// Reducer Helper Functions
-// =============================================================================
-
-/**
- * Convert a PendingInteractiveRequest from the backend to a PendingRequest for UI state.
- */
-function convertPendingRequest(req: PendingInteractiveRequest | null | undefined): PendingRequest {
-  if (!req) {
-    return { type: 'none' };
-  }
-
-  if (req.toolName === 'AskUserQuestion') {
-    const input = req.input as { questions?: unknown[] };
-    return {
-      type: 'question',
-      request: {
-        requestId: req.requestId,
-        questions: (input.questions ?? []) as UserQuestionRequest['questions'],
-        timestamp: req.timestamp,
-      },
-    };
-  }
-
-  return {
-    type: 'permission',
-    request: {
-      requestId: req.requestId,
-      toolName: req.toolName,
-      toolInput: req.input,
-      timestamp: req.timestamp,
-      planContent: req.planContent,
-    },
-  };
-}
-
-/**
- * Handle WS_CLAUDE_MESSAGE action - processes Claude messages and stores them.
- */
-function handleClaudeMessage(state: ChatState, claudeMsg: ClaudeMessage, order: number): ChatState {
-  // Transition from starting to running when receiving a Claude message
-  let baseState: ChatState =
-    state.sessionStatus.phase === 'starting'
-      ? { ...state, sessionStatus: { phase: 'running' } }
-      : state;
-
-  // Set to ready when we receive a result, and accumulate token stats
-  if (claudeMsg.type === 'result') {
-    baseState = {
-      ...baseState,
-      sessionStatus: { phase: 'ready' },
-      tokenStats: updateTokenStatsFromResult(baseState.tokenStats, claudeMsg),
-    };
-  }
-
-  // Check if message should be stored
-  if (!shouldStoreMessage(claudeMsg)) {
-    return baseState;
-  }
-
-  // Create and add the message using order-based insertion
-  const chatMessage = createClaudeMessage(claudeMsg, order);
-  const newMessages = insertMessageByOrder(baseState.messages, chatMessage);
-  const newIndex = newMessages.indexOf(chatMessage);
-
-  // Track tool_use message index for O(1) updates
-  const toolUseId = getToolUseIdFromMessage(claudeMsg);
-  if (toolUseId) {
-    const newToolUseIdToIndex = new Map(baseState.toolUseIdToIndex);
-    newToolUseIdToIndex.set(toolUseId, newIndex);
-    return { ...baseState, messages: newMessages, toolUseIdToIndex: newToolUseIdToIndex };
-  }
-
-  return { ...baseState, messages: newMessages };
-}
-
-/**
- * Handle TOOL_INPUT_UPDATE action - updates tool input with O(1) lookup.
- */
-function handleToolInputUpdate(
-  state: ChatState,
-  toolUseId: string,
-  input: Record<string, unknown>
-): ChatState {
-  // Try O(1) lookup first
-  let messageIndex = state.toolUseIdToIndex.get(toolUseId);
-  let currentState = state;
-  let needsIndexUpdate = false;
-
-  // If cached index exists, verify it points to the correct message
-  // (index may be stale if messages were inserted in the middle of the array)
-  if (messageIndex !== undefined) {
-    const cachedMsg = state.messages[messageIndex];
-    if (!isToolUseMessageWithId(cachedMsg, toolUseId)) {
-      // Cached index is stale, need to do linear scan
-      messageIndex = undefined;
-      needsIndexUpdate = true;
-    }
-  }
-
-  // Fallback to linear scan if not found or stale
-  if (messageIndex === undefined) {
-    messageIndex = state.messages.findIndex((msg) => isToolUseMessageWithId(msg, toolUseId));
-    if (messageIndex === -1) {
-      return state; // Tool use not found
-    }
-    needsIndexUpdate = true;
-  }
-
-  // Update index for future lookups if needed
-  if (needsIndexUpdate) {
-    const newToolUseIdToIndex = new Map(state.toolUseIdToIndex);
-    newToolUseIdToIndex.set(toolUseId, messageIndex);
-    currentState = { ...state, toolUseIdToIndex: newToolUseIdToIndex };
-  }
-
-  const msg = currentState.messages[messageIndex];
-
-  // Update the message with new input
-  const claudeMsg = msg.message;
-  const event = claudeMsg?.event as
-    | { type: 'content_block_start'; content_block: { type: string; input?: unknown } }
-    | undefined;
-  if (!event?.content_block) {
-    return currentState;
-  }
-
-  const updatedEvent = {
-    ...event,
-    content_block: {
-      ...event.content_block,
-      input,
-    },
-  };
-
-  const updatedChatMessage: ChatMessage = {
-    ...msg,
-    message: { ...claudeMsg, event: updatedEvent } as ClaudeMessage,
-  };
-
-  const newMessages = [...currentState.messages];
-  newMessages[messageIndex] = updatedChatMessage;
-  return { ...currentState, messages: newMessages };
-}
-
-// =============================================================================
-// Reducer Slices
-// =============================================================================
-
-type ReducerSlice = (state: ChatState, action: ChatAction) => ChatState;
-
-function reduceSessionSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'WS_STATUS':
-      return {
-        ...state,
-        sessionStatus: action.payload.running ? { phase: 'running' } : { phase: 'ready' },
-      };
-    case 'WS_STARTING':
-      return { ...state, sessionStatus: { phase: 'starting' } };
-    case 'WS_STARTED':
-      return { ...state, sessionStatus: { phase: 'running' }, latestThinking: null };
-    case 'WS_STOPPED':
-      return {
-        ...state,
-        sessionStatus: { phase: 'ready' },
-        toolProgress: new Map(),
-        isCompacting: false,
-        activeHooks: new Map(),
-      };
-    case 'WS_SESSIONS':
-      return { ...state, availableSessions: action.payload.sessions };
-    case 'SESSION_SWITCH_START':
-      return {
-        ...state,
-        ...createSessionSwitchResetState(),
-      };
-    case 'SESSION_LOADING_START':
-      return { ...state, sessionStatus: { phase: 'loading' } };
-    case 'STOP_REQUESTED':
-      return { ...state, sessionStatus: { phase: 'stopping' } };
-    default:
-      return state;
-  }
-}
-
-function reduceRequestSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'WS_PERMISSION_REQUEST':
-      return { ...state, pendingRequest: { type: 'permission', request: action.payload } };
-    case 'WS_USER_QUESTION':
-      return { ...state, pendingRequest: { type: 'question', request: action.payload } };
-    case 'PERMISSION_RESPONSE': {
-      const shouldDisablePlanMode =
-        action.payload.allow &&
-        state.pendingRequest.type === 'permission' &&
-        state.pendingRequest.request.toolName === 'ExitPlanMode';
-      return {
-        ...state,
-        pendingRequest: { type: 'none' },
-        ...(shouldDisablePlanMode && {
-          chatSettings: { ...state.chatSettings, planModeEnabled: false },
-        }),
-      };
-    }
-    case 'QUESTION_RESPONSE':
-      return { ...state, pendingRequest: { type: 'none' } };
-    case 'WS_PERMISSION_CANCELLED': {
-      const currentRequestId =
-        state.pendingRequest.type === 'permission'
-          ? state.pendingRequest.request.requestId
-          : state.pendingRequest.type === 'question'
-            ? state.pendingRequest.request.requestId
-            : null;
-      if (currentRequestId === action.payload.requestId) {
-        return { ...state, pendingRequest: { type: 'none' } };
-      }
-      return state;
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceSettingsSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'UPDATE_SETTINGS':
-      return { ...state, chatSettings: { ...state.chatSettings, ...action.payload } };
-    case 'SET_SETTINGS':
-      return { ...state, chatSettings: action.payload };
-    default:
-      return state;
-  }
-}
-
-function reduceMessageTransportSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'WS_CLAUDE_MESSAGE':
-      return handleClaudeMessage(state, action.payload.message, action.payload.order);
-    case 'WS_ERROR': {
-      const maxOrder = state.messages.reduce((max, m) => Math.max(max, m.order), -1);
-      const errorMsg: ClaudeMessage = {
-        type: 'error',
-        error: action.payload.message,
-        timestamp: new Date().toISOString(),
-      };
-      const errorChatMessage: ChatMessage = {
-        id: generateMessageId(),
-        source: 'claude',
-        message: errorMsg,
-        timestamp: new Date().toISOString(),
-        order: maxOrder + 1,
-      };
-      return {
-        ...state,
-        messages: [...state.messages, errorChatMessage],
-      };
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceMessageQueueSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'USER_MESSAGE_SENT':
-      return { ...state, messages: [...state.messages, action.payload] };
-    case 'ADD_TO_QUEUE': {
-      const newQueuedMessages = new Map(state.queuedMessages);
-      newQueuedMessages.set(action.payload.id, action.payload);
-      return {
-        ...state,
-        queuedMessages: newQueuedMessages,
-      };
-    }
-    case 'MESSAGE_SENDING': {
-      const { id, text, attachments } = action.payload;
-      const newPendingMessages = new Map(state.pendingMessages);
-      newPendingMessages.set(id, { text, attachments });
-      return {
-        ...state,
-        pendingMessages: newPendingMessages,
-      };
-    }
-    case 'CLEAR_REJECTED_MESSAGE':
-      return {
-        ...state,
-        lastRejectedMessage: null,
-      };
-    case 'MESSAGE_USED_AS_RESPONSE': {
-      const pendingContent = state.pendingMessages.get(action.payload.id);
-      const newPendingMessages = new Map(state.pendingMessages);
-      newPendingMessages.delete(action.payload.id);
-
-      if (state.messages.some((m) => m.id === action.payload.id)) {
-        return {
-          ...state,
-          pendingMessages: newPendingMessages,
-          pendingRequest: { type: 'none' },
-        };
-      }
-
-      const userMessage: ChatMessage = {
-        id: action.payload.id,
-        source: 'user',
-        text: action.payload.text,
-        timestamp: new Date().toISOString(),
-        attachments: pendingContent?.attachments,
-        order: action.payload.order,
-      };
-
-      return {
-        ...state,
-        messages: insertMessageByOrder(state.messages, userMessage),
-        pendingMessages: newPendingMessages,
-        pendingRequest: { type: 'none' },
-      };
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceMessageResetSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'CLEAR_CHAT': {
-      const sessionStatus: SessionStatus =
-        state.sessionStatus.phase === 'running' ? state.sessionStatus : { phase: 'ready' };
-      return {
-        ...state,
-        ...createBaseResetState(),
-        sessionStatus,
-        chatSettings: DEFAULT_CHAT_SETTINGS,
-      };
-    }
-    case 'RESET_FOR_SESSION_SWITCH':
-      return {
-        ...state,
-        ...createSessionSwitchResetState(),
-      };
-    default:
-      return state;
-  }
-}
-
-function reduceMessageSnapshotSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'MESSAGES_SNAPSHOT': {
-      const snapshotMessages = action.payload.messages;
-      const snapshotIds = new Set(snapshotMessages.map((m) => m.id));
-      const newPendingMessages = new Map<string, PendingMessageContent>();
-      for (const [id, content] of state.pendingMessages) {
-        if (!snapshotIds.has(id)) {
-          newPendingMessages.set(id, content);
-        }
-      }
-
-      const sessionStatus: SessionStatus = action.payload.sessionStatus;
-      const pendingRequest = convertPendingRequest(action.payload.pendingInteractiveRequest);
-
-      return {
-        ...state,
-        messages: snapshotMessages,
-        queuedMessages: new Map(),
-        sessionStatus,
-        pendingRequest,
-        toolUseIdToIndex: new Map(),
-        pendingMessages: newPendingMessages,
-        lastRejectedMessage: null,
-        messageIdToUuid: new Map(),
-        pendingUserMessageUuids: [],
-        localUserMessageIds: new Set(),
-        rewindPreview: null,
-      };
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceMessageStateMachineSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'MESSAGE_STATE_CHANGED':
-      return applyMessageStateChange(state, action.payload);
-    default:
-      return state;
-  }
-}
-
-function applyMessageStateChange(
-  state: ChatState,
-  payload: Extract<ChatAction, { type: 'MESSAGE_STATE_CHANGED' }>['payload']
-): ChatState {
-  const { id, newState, userMessage, errorMessage } = payload;
-
-  switch (newState) {
-    case MessageState.ACCEPTED:
-      return userMessage ? handleAcceptedState(state, id, userMessage) : state;
-    case MessageState.DISPATCHED:
-    case MessageState.COMMITTED:
-    case MessageState.COMPLETE:
-      return handleRemoveFromQueue(state, id);
-    case MessageState.CANCELLED:
-      return handleCancelledState(state, id);
-    case MessageState.REJECTED:
-    case MessageState.FAILED:
-      return handleRejectedOrFailedState(state, id, errorMessage);
-    default:
-      debug.log(`[chat-reducer] Ignoring state transition to ${newState} for message ${id}`);
-      return state;
-  }
-}
-
-function handleAcceptedState(
-  state: ChatState,
-  id: string,
-  userMessage: Extract<
-    Extract<ChatAction, { type: 'MESSAGE_STATE_CHANGED' }>['payload']['userMessage'],
-    object
-  >
-): ChatState {
-  const newPendingMessages = new Map(state.pendingMessages);
-  newPendingMessages.delete(id);
-
-  if (state.messages.some((m) => m.id === id)) {
-    const newQueuedMessages = new Map(state.queuedMessages);
-    newQueuedMessages.set(id, {
-      id,
-      text: userMessage.text,
-      timestamp: userMessage.timestamp,
-      attachments: userMessage.attachments,
-      settings: userMessage.settings ?? {
-        selectedModel: null,
-        thinkingEnabled: false,
-        planModeEnabled: false,
-      },
-    });
-    return {
-      ...state,
-      queuedMessages: newQueuedMessages,
-      pendingMessages: newPendingMessages,
-    };
-  }
-
-  const newMessage: ChatMessage = {
-    id,
-    source: 'user',
-    text: userMessage.text,
-    timestamp: userMessage.timestamp,
-    attachments: userMessage.attachments,
-    order: userMessage.order,
-  };
-
-  const newQueuedMessages = new Map(state.queuedMessages);
-  newQueuedMessages.set(id, {
-    id,
-    text: userMessage.text,
-    timestamp: userMessage.timestamp,
-    attachments: userMessage.attachments,
-    settings: userMessage.settings ?? {
-      selectedModel: null,
-      thinkingEnabled: false,
-      planModeEnabled: false,
-    },
-  });
-
-  const newLocalUserMessageIds = new Set(state.localUserMessageIds);
-  newLocalUserMessageIds.add(id);
-
-  let newMessageIdToUuid = state.messageIdToUuid;
-  let newPendingUuids = state.pendingUserMessageUuids;
-  if (state.pendingUserMessageUuids.length > 0) {
-    const [uuid, ...remainingUuids] = state.pendingUserMessageUuids;
-    newMessageIdToUuid = new Map(state.messageIdToUuid);
-    newMessageIdToUuid.set(id, uuid);
-    newPendingUuids = remainingUuids;
-  }
-
-  return {
-    ...state,
-    messages: insertMessageByOrder(state.messages, newMessage),
-    queuedMessages: newQueuedMessages,
-    pendingMessages: newPendingMessages,
-    messageIdToUuid: newMessageIdToUuid,
-    pendingUserMessageUuids: newPendingUuids,
-    localUserMessageIds: newLocalUserMessageIds,
-  };
-}
-
-function handleRemoveFromQueue(state: ChatState, id: string): ChatState {
-  const newQueuedMessages = new Map(state.queuedMessages);
-  newQueuedMessages.delete(id);
-  return { ...state, queuedMessages: newQueuedMessages };
-}
-
-function handleCancelledState(state: ChatState, id: string): ChatState {
-  const newQueuedMessages = new Map(state.queuedMessages);
-  newQueuedMessages.delete(id);
-  return {
-    ...state,
-    messages: state.messages.filter((m) => m.id !== id),
-    queuedMessages: newQueuedMessages,
-  };
-}
-
-function handleRejectedOrFailedState(
-  state: ChatState,
-  id: string,
-  errorMessage?: string
-): ChatState {
-  const queuedMessage = state.queuedMessages.get(id);
-  const pendingContent = state.pendingMessages.get(id);
-  const recoveryContent = queuedMessage ?? pendingContent;
-
-  const newQueuedMessages = new Map(state.queuedMessages);
-  newQueuedMessages.delete(id);
-
-  const newPendingMessages = new Map(state.pendingMessages);
-  newPendingMessages.delete(id);
-
-  return {
-    ...state,
-    messages: state.messages.filter((m) => m.id !== id),
-    queuedMessages: newQueuedMessages,
-    pendingMessages: newPendingMessages,
-    lastRejectedMessage: recoveryContent
-      ? {
-          text: queuedMessage?.text ?? pendingContent?.text ?? '',
-          attachments: recoveryContent.attachments,
-          error: errorMessage ?? 'Message failed',
-        }
-      : null,
-  };
-}
-
-function reduceMessageUuidSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'USER_MESSAGE_UUID_RECEIVED': {
-      const unmappedUserMessages = state.messages.filter(
-        (m) =>
-          m.source === 'user' &&
-          state.localUserMessageIds.has(m.id) &&
-          !state.messageIdToUuid.has(m.id)
-      );
-
-      if (unmappedUserMessages.length > 0) {
-        const targetMessage = unmappedUserMessages[0];
-        const newMap = new Map(state.messageIdToUuid);
-        newMap.set(targetMessage.id, action.payload.uuid);
-        return {
-          ...state,
-          messageIdToUuid: newMap,
-        };
-      }
-
-      debug.log(
-        `[chat-reducer] UUID received but no unmapped user message found, queueing: ${action.payload.uuid}`
-      );
-      return {
-        ...state,
-        pendingUserMessageUuids: [...state.pendingUserMessageUuids, action.payload.uuid],
-      };
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceMessageCompactSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'COMPACT_BOUNDARY': {
-      const maxOrder = state.messages.reduce((max, m) => Math.max(max, m.order), -1);
-      const compactBoundaryMessage: ChatMessage = {
-        id: `compact-boundary-${Date.now()}`,
-        source: 'claude',
-        message: {
-          type: 'system',
-          subtype: 'compact_boundary',
-        } as ClaudeMessage,
-        timestamp: new Date().toISOString(),
-        order: maxOrder + 1,
-      };
-      return {
-        ...state,
-        hasCompactBoundary: true,
-        messages: [...state.messages, compactBoundaryMessage],
-      };
-    }
-    default:
-      return state;
-  }
-}
-
-function reduceToolingSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'TOOL_INPUT_UPDATE':
-      return handleToolInputUpdate(state, action.payload.toolUseId, action.payload.input);
-    case 'TOOL_USE_INDEXED': {
-      const newToolUseIdToIndex = new Map(state.toolUseIdToIndex);
-      newToolUseIdToIndex.set(action.payload.toolUseId, action.payload.index);
-      return { ...state, toolUseIdToIndex: newToolUseIdToIndex };
-    }
-    case 'THINKING_DELTA':
-      return {
-        ...state,
-        latestThinking: (state.latestThinking ?? '') + action.payload.thinking,
-      };
-    case 'THINKING_CLEAR':
-      return { ...state, latestThinking: null };
-    case 'SDK_TOOL_PROGRESS': {
-      const { toolUseId, toolName, elapsedSeconds } = action.payload;
-      const newToolProgress = new Map(state.toolProgress);
-      newToolProgress.set(toolUseId, { toolName, elapsedSeconds });
-      return { ...state, toolProgress: newToolProgress };
-    }
-    case 'SDK_TOOL_USE_SUMMARY': {
-      const { precedingToolUseIds } = action.payload;
-      const newToolProgress = new Map(state.toolProgress);
-      for (const toolUseId of precedingToolUseIds) {
-        newToolProgress.delete(toolUseId);
-      }
-      return { ...state, toolProgress: newToolProgress };
-    }
-    case 'SDK_COMPACTING_START':
-      return { ...state, isCompacting: true };
-    case 'SDK_COMPACTING_END':
-      return { ...state, isCompacting: false };
-    default:
-      return state;
-  }
-}
-
-function reduceSystemSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'SDK_STATUS_UPDATE': {
-      const { permissionMode } = action.payload;
-      if (permissionMode !== undefined) {
-        return { ...state, permissionMode };
-      }
-      return state;
-    }
-    case 'SDK_TASK_NOTIFICATION': {
-      const newNotification: TaskNotification = {
-        id: crypto.randomUUID(),
-        message: action.payload.message,
-        timestamp: new Date().toISOString(),
-      };
-      return {
-        ...state,
-        taskNotifications: [...state.taskNotifications, newNotification],
-      };
-    }
-    case 'SYSTEM_INIT':
-      return { ...state, sessionInitData: action.payload, slashCommandsLoaded: true };
-    case 'HOOK_STARTED': {
-      const { hookId, hookName, hookEvent } = action.payload;
-      const newActiveHooks = new Map(state.activeHooks);
-      newActiveHooks.set(hookId, {
-        hookId,
-        hookName,
-        hookEvent,
-        startedAt: new Date().toISOString(),
-      });
-      return { ...state, activeHooks: newActiveHooks };
-    }
-    case 'HOOK_RESPONSE': {
-      const newActiveHooks = new Map(state.activeHooks);
-      newActiveHooks.delete(action.payload.hookId);
-      return { ...state, activeHooks: newActiveHooks };
-    }
-    case 'WS_SLASH_COMMANDS':
-      return {
-        ...state,
-        slashCommands: action.payload.commands,
-        slashCommandsLoaded: true,
-      };
-    case 'DISMISS_TASK_NOTIFICATION':
-      return {
-        ...state,
-        taskNotifications: state.taskNotifications.filter(
-          (notif) => notif.id !== action.payload.id
-        ),
-      };
-    case 'CLEAR_TASK_NOTIFICATIONS':
-      return { ...state, taskNotifications: [] };
-    default:
-      return state;
-  }
-}
-
-function reduceRewindPreviewSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'REWIND_PREVIEW_START':
-      return {
-        ...state,
-        rewindPreview: {
-          userMessageId: action.payload.userMessageId,
-          requestNonce: action.payload.requestNonce,
-          isLoading: true,
-        },
-      };
-    case 'REWIND_PREVIEW_SUCCESS': {
-      if (!state.rewindPreview) {
-        return state;
-      }
-      if (
-        action.payload.userMessageId &&
-        action.payload.userMessageId !== state.rewindPreview.userMessageId
-      ) {
-        return state;
-      }
-      return {
-        ...state,
-        rewindPreview: {
-          ...state.rewindPreview,
-          isLoading: false,
-          affectedFiles: action.payload.affectedFiles,
-        },
-      };
-    }
-    case 'REWIND_PREVIEW_ERROR': {
-      if (!state.rewindPreview) {
-        return state;
-      }
-      if (
-        action.payload.requestNonce &&
-        action.payload.requestNonce !== state.rewindPreview.requestNonce
-      ) {
-        return state;
-      }
-      if (
-        action.payload.userMessageId &&
-        action.payload.userMessageId !== state.rewindPreview.userMessageId
-      ) {
-        return state;
-      }
-      return {
-        ...state,
-        rewindPreview: {
-          ...state.rewindPreview,
-          isLoading: false,
-          error: action.payload.error,
-        },
-      };
-    }
-    case 'REWIND_CANCEL':
-      return { ...state, rewindPreview: null };
-    default:
-      return state;
-  }
-}
-
-function reduceRewindExecutionSlice(state: ChatState, action: ChatAction): ChatState {
-  switch (action.type) {
-    case 'REWIND_EXECUTING':
-      return state.rewindPreview
-        ? {
-            ...state,
-            rewindPreview: {
-              ...state.rewindPreview,
-              isLoading: true,
-              isExecuting: true,
-            },
-          }
-        : state;
-    case 'REWIND_SUCCESS': {
-      if (!state.rewindPreview) {
-        return state;
-      }
-      if (
-        action.payload.userMessageId &&
-        action.payload.userMessageId !== state.rewindPreview.userMessageId
-      ) {
-        return state;
-      }
-      return { ...state, rewindPreview: null };
-    }
-    default:
-      return state;
-  }
-}
-
-// =============================================================================
-// Reducer
-// =============================================================================
-
-const chatReducerSlices: ReducerSlice[] = [
-  reduceSessionSlice,
-  reduceRequestSlice,
-  reduceSettingsSlice,
-  reduceMessageTransportSlice,
-  reduceMessageQueueSlice,
-  reduceMessageResetSlice,
-  reduceMessageSnapshotSlice,
-  reduceMessageStateMachineSlice,
-  reduceMessageUuidSlice,
-  reduceMessageCompactSlice,
-  reduceToolingSlice,
-  reduceSystemSlice,
-  reduceRewindPreviewSlice,
-  reduceRewindExecutionSlice,
-];
-
-export function chatReducer(state: ChatState, action: ChatAction): ChatState {
-  let nextState = state;
-  for (const reduce of chatReducerSlices) {
-    const updated = reduce(nextState, action);
-    if (updated !== nextState) {
-      nextState = updated;
-    }
-  }
-  return nextState;
-}
-
-// =============================================================================
-// Action Creators (for type-safe dispatch)
-// =============================================================================
-
-// Individual message type handlers for createActionFromWebSocketMessage
-
-function handleStatusMessage(data: WebSocketMessage): ChatAction {
-  return { type: 'WS_STATUS', payload: { running: data.running ?? false } };
-}
-
-function handleClaudeMessageAction(data: WebSocketMessage): ChatAction | null {
-  if (isWsClaudeMessage(data) && data.order !== undefined) {
-    return { type: 'WS_CLAUDE_MESSAGE', payload: { message: data.data, order: data.order } };
-  }
-  return null;
-}
-
-function handleErrorMessageAction(data: WebSocketMessage): ChatAction | null {
-  if (data.message) {
-    return { type: 'WS_ERROR', payload: { message: data.message } };
-  }
-  return null;
-}
-
-function handleSessionsMessage(data: WebSocketMessage): ChatAction | null {
-  if (data.sessions) {
-    return { type: 'WS_SESSIONS', payload: { sessions: data.sessions } };
-  }
-  return null;
-}
-
-function handlePermissionRequestMessage(data: WebSocketMessage): ChatAction | null {
-  if (data.requestId && data.toolName) {
-    return {
-      type: 'WS_PERMISSION_REQUEST',
-      payload: {
-        requestId: data.requestId,
-        toolName: data.toolName,
-        toolInput: data.toolInput ?? {},
-        timestamp: new Date().toISOString(),
-        // Include plan content for ExitPlanMode requests
-        planContent: data.planContent ?? null,
-      },
-    };
-  }
-  return null;
-}
-
-function handleUserQuestionMessage(data: WebSocketMessage): ChatAction | null {
-  if (data.requestId && data.questions) {
-    return {
-      type: 'WS_USER_QUESTION',
-      payload: {
-        requestId: data.requestId,
-        questions: data.questions,
-        timestamp: new Date().toISOString(),
-      },
-    };
-  }
-  return null;
-}
-
-function handleMessagesSnapshot(data: WebSocketMessage): ChatAction | null {
-  if (!data.messages) {
-    return null;
-  }
-  return {
-    type: 'MESSAGES_SNAPSHOT',
-    payload: {
-      messages: data.messages,
-      sessionStatus: data.sessionStatus ?? { phase: 'ready' },
-      pendingInteractiveRequest: data.pendingInteractiveRequest ?? null,
-    },
-  };
-}
-
-function handleMessageStateChanged(data: WebSocketMessage): ChatAction | null {
-  if (!(data.id && data.newState)) {
-    return null;
-  }
-  // If userMessage exists, order is required
-  const userMessage =
-    data.userMessage && data.userMessage.order !== undefined
-      ? { ...data.userMessage, order: data.userMessage.order }
-      : undefined;
-  return {
-    type: 'MESSAGE_STATE_CHANGED',
-    payload: {
-      id: data.id,
-      newState: data.newState,
-      queuePosition: data.queuePosition,
-      errorMessage: data.errorMessage,
-      userMessage,
-    },
-  };
-}
-
-function handleToolProgressMessage(data: WebSocketMessage): ChatAction | null {
-  if (!(data.tool_use_id && data.tool_name && data.elapsed_time_seconds !== undefined)) {
-    return null;
-  }
-  return {
-    type: 'SDK_TOOL_PROGRESS',
-    payload: {
-      toolUseId: data.tool_use_id,
-      toolName: data.tool_name,
-      elapsedSeconds: data.elapsed_time_seconds,
-    },
-  };
-}
-
-function handleToolUseSummaryMessage(data: WebSocketMessage): ChatAction | null {
-  // Only require preceding_tool_use_ids (used for cleanup). Summary can be empty string.
-  if (!Array.isArray(data.preceding_tool_use_ids)) {
-    return null;
-  }
-  return {
-    type: 'SDK_TOOL_USE_SUMMARY',
-    payload: {
-      summary: data.summary,
-      precedingToolUseIds: data.preceding_tool_use_ids,
-    },
-  };
-}
-
-function handleSystemInitMessage(data: WebSocketMessage): ChatAction | null {
-  const initData = data.data as
-    | {
-        tools?: Array<{
-          name: string;
-          description?: string;
-          input_schema?: Record<string, unknown>;
-        }>;
-        model?: string;
-        cwd?: string;
-        apiKeySource?: string;
-        slashCommands?: string[];
-        plugins?: Array<{ name: string; path: string }>;
-      }
-    | undefined;
-  if (!initData) {
-    return null;
-  }
-  return {
-    type: 'SYSTEM_INIT',
-    payload: {
-      tools: initData.tools ?? [],
-      model: initData.model ?? null,
-      cwd: initData.cwd ?? null,
-      apiKeySource: initData.apiKeySource ?? null,
-      slashCommands: initData.slashCommands ?? [],
-      plugins: initData.plugins ?? [],
-    },
-  };
-}
-
-function handleHookStartedMessage(data: WebSocketMessage): ChatAction | null {
-  const hookData = data.data as
-    | {
-        hookId?: string;
-        hookName?: string;
-        hookEvent?: string;
-      }
-    | undefined;
-  if (!(hookData?.hookId && hookData?.hookName && hookData?.hookEvent)) {
-    return null;
-  }
-  return {
-    type: 'HOOK_STARTED',
-    payload: {
-      hookId: hookData.hookId,
-      hookName: hookData.hookName,
-      hookEvent: hookData.hookEvent,
-    },
-  };
-}
-
-function handleHookResponseMessage(data: WebSocketMessage): ChatAction | null {
-  const hookRespData = data.data as { hookId?: string } | undefined;
-  if (!hookRespData?.hookId) {
-    return null;
-  }
-  return { type: 'HOOK_RESPONSE', payload: { hookId: hookRespData.hookId } };
-}
-
-/**
- * Creates a ChatAction from a WebSocketMessage.
- * Returns null if the message type is not handled.
- */
-// biome-ignore lint/complexity/noExcessiveCognitiveComplexity: switch statement complexity is inherent to message type dispatch
-export function createActionFromWebSocketMessage(data: WebSocketMessage): ChatAction | null {
-  switch (data.type) {
-    case 'status':
-      return handleStatusMessage(data);
-    case 'starting':
-      return { type: 'WS_STARTING' };
-    case 'started':
-      return { type: 'WS_STARTED' };
-    case 'stopped':
-    case 'process_exit':
-      return { type: 'WS_STOPPED' };
-    case 'claude_message':
-      return handleClaudeMessageAction(data);
-    case 'error':
-      return handleErrorMessageAction(data);
-    case 'sessions':
-      return handleSessionsMessage(data);
-    case 'permission_request':
-      return handlePermissionRequestMessage(data);
-    case 'user_question':
-      return handleUserQuestionMessage(data);
-    // Request cancelled by CLI (e.g., Ctrl+C during permission or question prompt)
-    case 'permission_cancelled':
-      if (!data.requestId) {
-        // biome-ignore lint/suspicious/noConsole: Error logging for debugging malformed WebSocket messages
-        console.error('[Chat Reducer] Received permission_cancelled without requestId', data);
-      }
-      return { type: 'WS_PERMISSION_CANCELLED', payload: { requestId: data.requestId ?? '' } };
-    // Interactive response handling
-    case 'message_used_as_response':
-      return data.id && data.text && data.order !== undefined
-        ? {
-            type: 'MESSAGE_USED_AS_RESPONSE',
-            payload: { id: data.id, text: data.text, order: data.order },
-          }
-        : null;
-    // Message state machine events (primary protocol)
-    case 'messages_snapshot':
-      return handleMessagesSnapshot(data);
-    case 'message_state_changed':
-      return handleMessageStateChanged(data);
-    // SDK message type events
-    case 'tool_progress':
-      return handleToolProgressMessage(data);
-    case 'tool_use_summary':
-      return handleToolUseSummaryMessage(data);
-    case 'status_update':
-      return {
-        type: 'SDK_STATUS_UPDATE',
-        payload: { permissionMode: data.permissionMode },
-      };
-    case 'task_notification':
-      return data.message
-        ? { type: 'SDK_TASK_NOTIFICATION', payload: { message: data.message } }
-        : null;
-    // System subtype events
-    case 'system_init':
-      return handleSystemInitMessage(data);
-    case 'compact_boundary':
-      return { type: 'COMPACT_BOUNDARY' };
-    case 'hook_started':
-      return handleHookStartedMessage(data);
-    case 'hook_response':
-      return handleHookResponseMessage(data);
-    // Context compaction events
-    case 'compacting_start':
-      return { type: 'SDK_COMPACTING_START' };
-    case 'compacting_end':
-      return { type: 'SDK_COMPACTING_END' };
-    // Slash commands discovery
-    case 'slash_commands':
-      return data.slashCommands
-        ? { type: 'WS_SLASH_COMMANDS', payload: { commands: data.slashCommands } }
-        : null;
-    // User message UUID tracking (for rewind functionality)
-    case 'user_message_uuid':
-      return data.uuid
-        ? { type: 'USER_MESSAGE_UUID_RECEIVED', payload: { uuid: data.uuid } }
-        : null;
-    // Rewind files response events
-    case 'rewind_files_preview':
-      // If dryRun is false, this is the actual rewind completion
-      if (data.dryRun === false) {
-        return { type: 'REWIND_SUCCESS', payload: { userMessageId: data.userMessageId } };
-      }
-      // Otherwise, this is a preview (dry run) response
-      return {
-        type: 'REWIND_PREVIEW_SUCCESS',
-        payload: {
-          affectedFiles: data.affectedFiles ?? [],
-          userMessageId: data.userMessageId,
-        },
-      };
-    case 'rewind_files_error':
-      return data.rewindError
-        ? {
-            type: 'REWIND_PREVIEW_ERROR',
-            payload: { error: data.rewindError, userMessageId: data.userMessageId },
-          }
-        : null;
-    default:
-      return null;
-  }
-}
-
-/**
- * Creates a user message action.
- */
-export function createUserMessageAction(text: string, order: number): ChatAction {
-  const chatMessage: ChatMessage = {
-    id: generateMessageId(),
-    source: 'user',
-    text,
-    timestamp: new Date().toISOString(),
-    order,
-  };
-  return { type: 'USER_MESSAGE_SENT', payload: chatMessage };
-}
+export * from './reducer';
