+++ src/components/chat/reducer/slices/tooling.ts
@@ -1,0 +1,41 @@
+import { handleToolInputUpdate } from '../helpers';
+import type { ChatAction, ChatState } from '../types';
+
+export function reduceToolingSlice(state: ChatState, action: ChatAction): ChatState {
+  switch (action.type) {
+    case 'TOOL_INPUT_UPDATE':
+      return handleToolInputUpdate(state, action.payload.toolUseId, action.payload.input);
+    case 'TOOL_USE_INDEXED': {
+      const newToolUseIdToIndex = new Map(state.toolUseIdToIndex);
+      newToolUseIdToIndex.set(action.payload.toolUseId, action.payload.index);
+      return { ...state, toolUseIdToIndex: newToolUseIdToIndex };
+    }
+    case 'THINKING_DELTA':
+      return {
+        ...state,
+        latestThinking: (state.latestThinking ?? '') + action.payload.thinking,
+      };
+    case 'THINKING_CLEAR':
+      return { ...state, latestThinking: null };
+    case 'SDK_TOOL_PROGRESS': {
+      const { toolUseId, toolName, elapsedSeconds } = action.payload;
+      const newToolProgress = new Map(state.toolProgress);
+      newToolProgress.set(toolUseId, { toolName, elapsedSeconds });
+      return { ...state, toolProgress: newToolProgress };
+    }
+    case 'SDK_TOOL_USE_SUMMARY': {
+      const { precedingToolUseIds } = action.payload;
+      const newToolProgress = new Map(state.toolProgress);
+      for (const toolUseId of precedingToolUseIds) {
+        newToolProgress.delete(toolUseId);
+      }
+      return { ...state, toolProgress: newToolProgress };
+    }
+    case 'SDK_COMPACTING_START':
+      return { ...state, isCompacting: true };
+    case 'SDK_COMPACTING_END':
+      return { ...state, isCompacting: false };
+    default:
+      return state;
+  }
+}
